+++ src/program_data_api/fetch_standard_deductions.js
@@ -14,7 +14,7 @@ export class FetchStandardDeductions {
         const NON_DEFAULT_STATES_TERRITORIES = [
             'AK',
             'HI',
-            'IL',
+            // 'IL', Just use defaults for now
             'GU',
             'VI',
         ];